LayoutTests/media/picture-in-picture-api-exit-pip-2-expected.txt

This tests that exit Picture-in-Picture rejects when there is no Picture-in-Picture video.

EXPECTED (error.name == 'InvalidStateError') OK
END OF TEST
